remote-tmux: split proxy retry tests under budget
