+++ cmuxTests/RemoteTmuxAuthTests.swift
@@ -67,134 +67,6 @@ import Testing
         #expect(RemoteTmuxSSHTransport.indicatesAuthRequired(stderr))
     }
 
-    // MARK: - Proxy-closed transport classification
-
-    /// SILENT proxy closures — the placeholder is the only diagnostic in
-    /// stderr — fire the predicate. These are the cases an interactive retry
-    /// recovers because the proxy's own auth/2FA prompt was suppressed under
-    /// BatchMode and a tty lets it surface. One mixed-case fixture guards the
-    /// `lowercased()` path against a future literal-only refactor; OpenSSH
-    /// itself emits the placeholder in caps.
-    @Test(arguments: [
-        "Connection closed by UNKNOWN port 65535",
-        "ssh_dispatch_run_fatal: Connection to UNKNOWN port 65535: Broken pipe",
-        "Connection closed by UnKnOwN port 65535",
-    ])
-    func classifiesSilentProxyCommandClosures(_ stderr: String) {
-        #expect(RemoteTmuxSSHTransport.indicatesProxyCommandTransportClosed(stderr))
-    }
-
-    /// The placeholder also stamps stderr when the proxy explains *why* it
-    /// closed: target unreachable behind a ProxyJump, `nc` to a refused port,
-    /// stdio forwarding teardown, target spoke no SSH on the negotiated port,
-    /// etc. None of those are recoverable by re-running ssh under a tty, so
-    /// the predicate MUST return false when any diagnostic marker is present
-    /// alongside the placeholder — otherwise the controller bounces the user
-    /// through a futile interactive prompt instead of surfacing the real
-    /// proxy error.
-    @Test(arguments: [
-        "channel 0: open failed: connect failed: Connection refused\nstdio forwarding failed\nConnection closed by UNKNOWN port 65535",
-        "connect failed: Connection refused\nConnection closed by UNKNOWN port 65535",
-        "stdio forwarding failed\nssh_exchange_identification: Connection closed by remote host\nConnection closed by UNKNOWN port 65535",
-        "kex_exchange_identification: Connection closed by remote host\nConnection closed by UNKNOWN port 65535",
-        "ssh: Could not resolve hostname inner.invalid: nodename nor servname provided\nConnection closed by UNKNOWN port 65535",
-        // BSD/macOS `nc` ProxyCommand: nc's own getaddrinfo error propagates
-        // raw without OpenSSH's `Could not resolve hostname` wrapper.
-        "nc: getaddrinfo: nodename nor servname provided, or not known\nConnection closed by UNKNOWN port 65535",
-        "channel 1: open failed: administratively prohibited: open failed\nConnection closed by UNKNOWN port 65535",
-        // Linux `nc` ProxyCommand TCP connect timeout — phrased "Connection timed
-        // out" (not BSD/macOS's "Operation timed out"), with no other marker.
-        "nc: connect to inner.invalid port 22 (tcp) failed: Connection timed out\nConnection closed by UNKNOWN port 65535",
-        // Banner-exchange closure standing alone (no co-present marker): the inner
-        // target dropped us pre-auth (fail2ban / tcpwrappers / not-SSH-on-port).
-        "ssh_exchange_identification: Connection closed by remote host\nConnection closed by UNKNOWN port 65535",
-    ])
-    func doesNotClassifyExplainedProxyClosures(_ stderr: String) {
-        #expect(!RemoteTmuxSSHTransport.indicatesProxyCommandTransportClosed(stderr))
-    }
-
-    /// The match is anchored to OpenSSH's exact phrasings (`to UNKNOWN port
-    /// 65535` / `by UNKNOWN port 65535`) so remote-stderr noise that merely
-    /// mentions the words in a different order — `pam_motd`, `~/.ssh/rc`, or a
-    /// remote command spelling out the literal `UNKNOWN port 65535` without
-    /// OpenSSH's connection-formed preposition — does NOT misclassify into the
-    /// interactive-retry path.
-    @Test(arguments: [
-        "MOTD: lab name is UNKNOWN port 65535 status board",
-        "remote warning: process listening on port 65535 with unknown owner",
-        "user note: 'unknown port 65535' is reserved",
-    ])
-    func anchorsProxyClosedMatchToOpenSSHPhrasing(_ stderr: String) {
-        #expect(!RemoteTmuxSSHTransport.indicatesProxyCommandTransportClosed(stderr))
-    }
-
-    /// Real-port transport failures (a host that's actually unreachable on its
-    /// real port) MUST NOT be classified as proxy-closed — they will not be
-    /// fixed by an interactive retry and should surface as genuine errors
-    /// instead. Generic non-auth fixtures (empty stderr, unrelated text,
-    /// algorithm-negotiation) are covered by ``doesNotClassifyNonAuthFailures``
-    /// for the sibling predicate; the cases here are predicate-specific
-    /// (proxy-shaped strings that must not collide with real-port stderr).
-    @Test(arguments: [
-        "ssh: connect to host bad.example.com port 22: Connection refused",
-        "ssh: connect to host bad.example.com port 2222: Operation timed out",
-        "Connection closed by 10.0.0.5 port 22",
-    ])
-    func doesNotClassifyRealPortClosuresAsProxyTransport(_ stderr: String) {
-        #expect(!RemoteTmuxSSHTransport.indicatesProxyCommandTransportClosed(stderr))
-    }
-
-    /// The proxy-closed and auth-required predicates are independent: the
-    /// proxy-closed stderr should NOT be treated as an explicit auth failure,
-    /// and a clean `Permission denied` should NOT trip the proxy-closed match.
-    /// The controller routes through the composed predicate
-    /// ``indicatesInteractiveRetryWillHelp`` precisely because they're
-    /// different signals that share a recovery (interactive ssh retry).
-    @Test func proxyClosedAndAuthRequiredAreDisjoint() {
-        let proxyOnly = "ssh_dispatch_run_fatal: Connection to UNKNOWN port 65535: Broken pipe"
-        #expect(RemoteTmuxSSHTransport.indicatesProxyCommandTransportClosed(proxyOnly))
-        #expect(!RemoteTmuxSSHTransport.indicatesAuthRequired(proxyOnly))
-
-        let authOnly = "user@host: Permission denied (publickey,password)."
-        #expect(RemoteTmuxSSHTransport.indicatesAuthRequired(authOnly))
-        #expect(!RemoteTmuxSSHTransport.indicatesProxyCommandTransportClosed(authOnly))
-    }
-
-    // MARK: - Composed routing predicate
-
-    /// ``indicatesInteractiveRetryWillHelp`` is what every controller catch
-    /// site routes through. It MUST fire for every stderr either sibling
-    /// predicate fires for, so a future signal added to one constituent (or a
-    /// new third constituent) propagates to all routing sites at once instead
-    /// of silently regressing whichever catch still spells out only one of
-    /// the two. Asserted both directions for each stderr style.
-    @Test(arguments: [
-        // Auth-required side
-        "user@host: Permission denied (publickey,password).",
-        "Host key verification failed.",
-        "Too many authentication failures",
-        // Proxy-closed side
-        "Connection closed by UNKNOWN port 65535",
-        "ssh_dispatch_run_fatal: Connection to UNKNOWN port 65535: Broken pipe",
-    ])
-    func composedPredicateFiresForEitherRecoverableSignal(_ stderr: String) {
-        #expect(RemoteTmuxSSHTransport.indicatesInteractiveRetryWillHelp(stderr))
-    }
-
-    @Test(arguments: [
-        "no server running on /tmp/tmux-501/default",
-        "no matching host key type found. their offer: ssh-rsa",
-        "ssh: connect to host bad.example.com port 22: Connection refused",
-        "",
-        // An explained proxy closure (target unreachable behind ProxyJump) must
-        // not slip through the composed predicate either — surface the real
-        // error instead of bouncing the user through interactive auth.
-        "channel 0: open failed: connect failed: Connection refused\nstdio forwarding failed\nConnection closed by UNKNOWN port 65535",
-    ])
-    func composedPredicateRejectsNonRecoverableFailures(_ stderr: String) {
-        #expect(!RemoteTmuxSSHTransport.indicatesInteractiveRetryWillHelp(stderr))
-    }
-
     @Test(arguments: [
         "command refresh-client: unknown flag -B",
         "refresh-client: unknown option -- B",
